pyproject.toml
[project]
name = "senior_ai_engineer_assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "Kiran Bagade"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "ipykernel (>=6.30.1,<7.0.0)",
    "llama-index-llms-google-genai (>=0.3.0,<0.4.0)",
    "llama-index (>=0.13.4,<0.14.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "mammoth (>=1.10.0,<2.0.0)",
    "docxtpl (>=0.20.1,<0.21.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
